requirements.txt
pip~=22.3.1
google~=3.0.0
requests~=2.26.0
aiogram~=2.25.1
python-dotenv~=0.18.0